STBPLATFORM-139: Implement "simple" variantPolicy - no nested variants.
This is necessary to interact with some applications, for example Connman's SetProperty(sv) methods expects the variant to resolve to a{ss}. The default policy would give a{sv}, with the nested variant resolving to string.

The default policy is much more flexible but some applications are too strict about parameter types, hence the need for "simple variant policy".